fix exceed length swiftlint warning
